+++ src/components/navigation/MonthYearDesktopNavigator.vue
